QuickAdminPanel automatic commit

+++ app/Http/Controllers/Api/V1/Admin/ProductTagApiController.php
@@ -5,34 +5,52 @@ namespace App\Http\Controllers\Api\V1\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreProductTagRequest;
 use App\Http\Requests\UpdateProductTagRequest;
+use App\Http\Resources\Admin\ProductTagResource;
 use App\ProductTag;
+use Gate;
+use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 class ProductTagApiController extends Controller
 {
     public function index()
     {
-        $productTags = ProductTag::all();
+        abort_if(Gate::denies('product_tag_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return $productTags;
+        return new ProductTagResource(ProductTag::all());
     }
 
     public function store(StoreProductTagRequest $request)
     {
-        return ProductTag::create($request->all());
-    }
+        $productTag = ProductTag::create($request->all());
 
-    public function update(UpdateProductTagRequest $request, ProductTag $productTag)
-    {
-        return $productTag->update($request->all());
+        return (new ProductTagResource($productTag))
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);
     }
 
     public function show(ProductTag $productTag)
     {
-        return $productTag;
+        abort_if(Gate::denies('product_tag_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        return new ProductTagResource($productTag);
+    }
+
+    public function update(UpdateProductTagRequest $request, ProductTag $productTag)
+    {
+        $productTag->update($request->all());
+
+        return (new ProductTagResource($productTag))
+            ->response()
+            ->setStatusCode(Response::HTTP_ACCEPTED);
     }
 
     public function destroy(ProductTag $productTag)
     {
-        return $productTag->delete();
+        abort_if(Gate::denies('product_tag_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $productTag->delete();
+
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
